Updated tailwind config
