+++ gatsby-theme-contentful/gatsby-browser.js
@@ -1,0 +1,1 @@
+import './src/styles/style.css';
