+++ gatsby-theme-contentful/package.json
@@ -5,7 +5,7 @@
     "type": "git",
     "url": "https://github.com/ethriel3695/gatsby-theme-contentful.git"
   },
-  "version": "0.0.52",
+  "version": "0.0.63",
   "main": "index.js",
   "license": "MIT",
   "peerDependencies": {
