+++ gatsby-theme-contentful/postcss.config.js
@@ -1,0 +1,3 @@
+module.exports = () => ({
+  plugins: [require('tailwindcss'), require(`autoprefixer`)()],
+});
