+++ gatsby-theme-contentful/src/hooks/slugList.js
@@ -3,6 +3,12 @@ import { graphql, useStaticQuery } from 'gatsby';
 export const useSlugList = () => {
   const data = useStaticQuery(graphql`
     query SlugList {
+      allContentfulPage(filter: { slug: { ne: "/placeholder-content" } }) {
+        nodes {
+          title
+          slug
+        }
+      }
       allMdx(
         sort: { order: DESC, fields: [frontmatter___date] }
         filter: { frontmatter: { published: { eq: true } } }
@@ -14,12 +20,6 @@ export const useSlugList = () => {
             slug
             label
           }
-        }
-      }
-      allContentfulPage {
-        nodes {
-          title
-          slug
         }
       }
     }
