+++ gatsby-theme-contentful/src/utils/buildNav.js
@@ -1,18 +1,18 @@
 export const buildNav = navList => {
   let navs = [];
   let navObject = null;
-  navList.allMdx.nodes.map(node => {
-    navObject = {
-      route: node.frontmatter.slug,
-      label: node.frontmatter.label,
-    };
-    navs.push(navObject);
-    return true;
-  });
   navList.allContentfulPage.nodes.map(node => {
     navObject = {
       route: node.slug,
       label: node.title,
+    };
+    navs.push(navObject);
+    return true;
+  });
+  navList.allMdx.nodes.map(node => {
+    navObject = {
+      route: node.frontmatter.slug,
+      label: node.frontmatter.label,
     };
     navs.push(navObject);
     return true;
