+++ gatsby-theme-contentful/src/utils/displayNotifications.js
@@ -1,0 +1,15 @@
+const displayNotification = () => {
+  if (Notification.permission === 'granted') {
+    navigator.serviceWorker.getRegistration().then(function(reg) {
+      var options = {
+        body: 'Here is a notification body!',
+        vibrate: [100, 50, 100],
+        data: {
+          dateOfArrival: Date.now(),
+          primaryKey: 1,
+        },
+      };
+      reg && reg.showNotification('Hello world!', options);
+    });
+  }
+};
