+++ gatsby-theme-contentful/tailwind.config.js
