--- postcss.config.js
@@ -1,4 +1,0 @@
-const tailwind = require('tailwindcss');
-module.exports = () => ({
-  plugins: [tailwind('../tailwind.config.js'), require(`autoprefixer`)()],
-});
